moving errors out of constants
