+++ demisto_sdk/commands/common/hook_validations/base_validator.py
@@ -4,7 +4,7 @@ import re
 from abc import abstractmethod
 
 import yaml
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.structure import \
     StructureValidator
 from demisto_sdk.commands.common.tools import (
