+++ demisto_sdk/commands/common/hook_validations/description.py
@@ -2,7 +2,8 @@ import glob
 
 from demisto_sdk.commands.common.constants import (BETA_INTEGRATION_DISCLAIMER,
                                                    BETA_INTEGRATION_REGEX,
-                                                   INTEGRATION_REGEX, Errors)
+                                                   INTEGRATION_REGEX)
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.tools import (get_yaml, os, print_error,
                                                print_warning, re)
