+++ demisto_sdk/commands/common/hook_validations/incident_field.py
@@ -5,7 +5,7 @@ import re
 from distutils.version import LooseVersion
 from enum import Enum, IntEnum
 
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
 from demisto_sdk.commands.common.tools import print_error
