+++ demisto_sdk/commands/common/hook_validations/layout.py
@@ -1,4 +1,4 @@
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
 from demisto_sdk.commands.common.tools import print_error
