+++ demisto_sdk/commands/common/hook_validations/old_release_notes.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 import os
 import re
 
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.tools import (
     old_get_latest_release_notes_text, old_get_release_notes_file_path,
     print_error, run_command)
