+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -11,7 +11,8 @@ from demisto_sdk.commands.common.constants import (  # PACK_METADATA_PRICE,
     PACK_METADATA_DESC, PACK_METADATA_FIELDS, PACK_METADATA_KEYWORDS,
     PACK_METADATA_NAME, PACK_METADATA_TAGS, PACK_METADATA_USE_CASES,
     PACKS_PACK_IGNORE_FILE_NAME, PACKS_PACK_META_FILE_NAME,
-    PACKS_README_FILE_NAME, PACKS_WHITELIST_FILE_NAME, Errors)
+    PACKS_README_FILE_NAME, PACKS_WHITELIST_FILE_NAME)
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.tools import pack_name_to_path
