+++ demisto_sdk/commands/common/hook_validations/playbook.py
@@ -1,7 +1,7 @@
 from typing import Dict
 
 import click
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
 from demisto_sdk.commands.common.tools import LOG_COLORS, print_error
