+++ demisto_sdk/commands/common/hook_validations/readme.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.tools import (get_content_path, print_error,
                                                print_warning, run_command_os)
