+++ demisto_sdk/commands/common/hook_validations/reputation.py
@@ -1,6 +1,6 @@
 from distutils.version import LooseVersion
 
-from demisto_sdk.commands.common.constants import Errors
+from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
 from demisto_sdk.commands.common.tools import print_error
